fix: change type to reduce warning

+++ src/main.cpp
@@ -78,7 +78,7 @@ int main() {
 
     // ==================== Construct and run model ==========================
 
-    for (long i {1}; i < Compartment::timesFollowUp; i++) {
+    for (size_t i {1}; i < Compartment::timesFollowUp; i++) {
         myModel.getModel()->update(i);
         // Debug: view each time step update
 //        viewModelUpdate(myModel.getModel(), i);
